KS-1-4 refactor

+++ src/main/kotlin/lesson_1/lesson1_task4.kt
@@ -1,10 +1,10 @@
 package org.example.lesson_1
 
 fun main() {
-    val length: ULong = 40868600000uL
-    val age: UByte = 27u
+    val length: Long = 40868600000L
+    val age: Byte = 27
     val partOfDay: Float = 0.075f
-    val seconds: UInt = 6480u
+    val seconds: Short = 6480
     val partOfYear: Double = 0.000205479453
     val apogee: Int = 327000
